requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
croniter==1.0.1
Django==3.1.5
djangorestframework==3.12.2
idna==2.10
natsort==7.1.0
python-dateutil==2.8.1
pytz==2020.5
redis==3.5.3
requests==2.25.1
rq==1.7.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
